pyproject.toml
[project]
name = "minigist"
description = "A tool that generates concise summaries for you Miniflux feeds."
readme = "README.md"
license-files = ["LICENSE"]
requires-python = ">=3.13"
dynamic = ["version"]
authors = [
    {name = "eikendev"},
]
maintainers = [
    {name = "eikendev"},
]
keywords = ["miniflux", "rss", "feed", "ai", "summarization", "cli"]
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
]
dependencies = [
    "apprise>=1.9.3",
    "click>=8.1.8",
    "markdown>=3.8",
    "miniflux>=1.1.3",
    "pydantic-ai>=0.1.6",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "structlog>=25.3.0",
    "trafilatura>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/eikendev/minigist"

[project.scripts]
minigist = "minigist.cli:cli"

[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["minigist"]

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "no-local-version"

[tool.ruff]
line-length = 120
lint.select = ["F", "E", "W", "B", "I", "SIM", "UP", "N", "FURB"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.11.7",
    "types-markdown>=3.8.0.20250415",
    "types-pyyaml>=6.0.12.20250402",
]
